pyproject.toml
[project]
name = "indekkusu"
version = "0.1.0"
description = "index and search doujinshi"
authors = [{ name = "Aloxaf", email = "aloxafx@gmail.com" }]
dependencies = [
    "click>=8.1.7",
    "opencv-python>=4.9.0.80",
    "tqdm>=4.66.4",
    "loguru>=0.7.2",
    "blake3>=0.4.1",
    "python-orb-slam3>=0.1.1",
    "peewee>=3.17.5",
    "faiss-cpu>=1.8.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "Apache-2.0" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
indekkusu = "indekkusu.cli:cli"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "prettytable>=3.10.0",
]
